Moving Battlecry handling to Minion class
Minion class now handles targetable battlecry (those battlecries for
which the user has to select the target).  The subclasses of Minion now
only has to worry about what happens to a single target.  This should
reduce some of the code bloat and result in simpler implementation of
the concrete minion classes.

For now, updating Alexstrasza for demonstration.  Complete update to
follow.