Added a utility for generating randomized decks
The DeckFactory class is a class that allows for various kinds of decks
to be randomly generated.  The factory is initialized by using a
builder so that various options can be utilized to filter on rarity,
character class, etc.

I also added an equals method and a hashCode method so that they were
consisted with the compare method which compares first on mana cost and
then on name (i.e. compareTo only returns 0 in the current
implementation if two cards have the same name).